Move general functions out of specific alg
